allow archiving even if android does not support it
the archived column will not be synchronized anymore (it still happens in messages->android direction if possible) if android is bugged or the archived column does not exist.

On my device (Samsung A22 5G) the archived column does not seem to get properly saved for all threads. I have no clue why, but every time it starts the app (or refreshes the message list) it reads a `false` for some messages from androids content resolver. The update statement affects 0 rows. I don't know exactly what is going on, but if that is the case it stores the data in the application database only. As a side effect it allows archiving messages even when the underlying os does not support it.